+++ src/main/java/blackjack/model/BlackjackGame.java
@@ -1,10 +1,9 @@
 package blackjack.model;
 
-import blackjack.ParticipantDto;
+import blackjack.dto.ParticipantDto;
 import blackjack.model.card.CardDeck;
 import blackjack.model.player.Dealer;
 import blackjack.model.player.Participant;
-import blackjack.model.player.Player;
 import blackjack.model.player.Players;
 import blackjack.model.result.BlackjackGameResult;
